scanner_integration/requirements.txt
absl-py==1.3.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
autopep8==2.0.1
backcall==0.2.0
bitarray==2.6.1
cachetools==5.2.1
camelot-py==0.10.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
distro==1.8.0
editdistance==0.6.2
efficientnet==1.0.0
entrypoints==0.4
essential-generators==1.0
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==1.2.0
flatbuffers==23.1.4
fonttools==4.38.0
gast==0.4.0
ghostscript==0.7
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
html5lib==1.1
idna==3.4
imageio==2.4.1
imgaug==0.4.0
importlib-metadata==6.0.0
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.0
keras==2.11.0
Keras-Applications==1.0.8
keras-ocr==0.9.2
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==3.0
numpy~=1.24.0
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
pdfminer.six==20221105
pdfreader==0.1.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow~=8.2.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py2exe==0.13.0.0
tk==0.1.0
future~=0.18.2
ttkthemes~=3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
Pygments==2.13.0
pyparsing==3.0.9
PyPDF2==2.12.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyWavelets==1.4.1
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.10.0
selenium==4.7.2
selenium-page-factory==2.5
shapely==2.0.0
signature-detect==0.1.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
stamp_processing==0.0.3
tabula-py==2.6.0
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
tifffile==2022.10.10
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
validators==0.20.0
wcwidth==0.2.5
webdriver-manager==3.8.5
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.11.0
pyzbar==0.1.9
tkcalendar==1.6.1
pytwain~=2.0